python/requirements.txt
python-dotenv
beautifulsoup4
selenium
discord
netcat